# src/main/resources/yapi_export_api/swaggerApi.json
{
  "swagger": "2.0",
  "info": {
    "title": "ERP-METRICS",
    "version": "last",
    "description": "erp-metrics 雷神架构组项目"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/app/monitor/config/update": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "修改应用配置",
        "description": "",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "raw",
            "in": "body",
            "description": "raw paramter",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n  \"id\":1,\n  \"appName\": \"crm异常监控\",\n  \"appOwner\": \"老吴\",\n  \"esIndex\": \"ld-erp-crm*\",\n  \"ddWebhookAccessToken\": \"d361cf7a46e47588395372602549149e243ac7fa8f1cc84e1e5eea3c339a7a47\",\n  \"esParam\": \"{\\\"query\\\":{\\\"bool\\\":{\\\"filter\\\":[{\\\"term\\\":{\\\"message\\\":\\\"error\\\"}},{\\\"range\\\":{\\\"index_utc_date\\\":{\\\"gte\\\":\\\"{startTime}\\\",\\\"lt\\\":\\\"{endTime}\\\"}}}],\\\"must_not\\\":[{\\\"term\\\":{\\\"log.file.path\\\":\\\"info.log\\\"}}]}}}\",\n  \"ddRobotMobile\": \"15338887312\",\n  \"alarmThreshold\": 1,\n  \"phoneAlarmThreshold\":10,\n  \"executeCycle\": 10,\n  \"deptId\":1,\n  \"testerName\":\"测试\",\n  \"testerPhone\":\"13888888888\"\n}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n    \"msg\": \"成功\",\n    \"result\": 1,\n    \"data\": true\n}"
            }
          }
        }
      }
    },
...
    "/app/monitor/config/detail": {
      "get": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "应用监控详情",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "配置主键id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n  \"msg\": \"成功\",\n  \"result\": 1,\n  \"data\": {\n    \"id\": \"1\",\n    \"appName\": \"雷神网关应用erp-gateway异常监控\",\n    \"appOwner\": \"\",\n    \"esIndex\": \"erp-gateway-*\",\n    \"esParam\": \"{\\\"query\\\":{\\\"bool\\\":{\\\"filter\\\":[{\\\"term\\\":{\\\"message\\\":\\\"Internal Server Error\\\"}},{\\\"range\\\":{\\\"index_utc_date\\\":{\\\"gte\\\":\\\"{startTime}\\\",\\\"lt\\\":\\\"{endTime}\\\"}}}],\\\"must_not\\\":[{\\\"term\\\":{\\\"log.file.path\\\":\\\"info.log\\\"}}]}}}\",\n    \"ddWebhookAccessToken\": \"d361cf7a46e47588395372602549149e243ac7fa8f1cc84e1e5eea3c339a7a47\",\n    \"ddRobotMobile\": \"15338887312\",\n    \"alarmThreshold\": 1,\n    \"phoneAlarmThreshold\":10,\n    \"lastScanTime\": \"2020-12-23 10:10:00\",\n    \"executeCycle\": 10,\n    \"deptId\":1,\n    \"deptName\":\"雷神项目部\",\n    \"testerName\":\"测试\",\n    \"testerPhone\":\"13888888888\"\n  }\n}"
            }
          }
        }
      }
    },
    "/app/monitor/config/configOwner": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "应用监控负责人列表",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "deptId",
            "in": "formData",
            "required": false,
            "description": "应用归属项目部的ID",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n\t\"msg\": \"成功\",\n\t\"result\": 1,\n\t\"data\": [{\n\t\t\"appOwner\": \"钦争艳\"\n\t}, {\n\t\t\"appOwner\": \"熊哲\"\n\t}, {\n\t\t\"appOwner\": \"胡云\"\n\t}, {\n\t\t\"appOwner\": \"江超\"\n\t}, {\n\t\t\"appOwner\": \"王国伟\"\n\t}, {\n\t\t\"appOwner\": \"姚宇明\"\n\t}, {\n\t\t\"appOwner\": \"郑韬\"\n\t}, {\n\t\t\"appOwner\": \"李坤保\"\n\t}, {\n\t\t\"appOwner\": \"罗玉华\"\n\t}, {\n\t\t\"appOwner\": \"黄兵\"\n\t}, {\n\t\t\"appOwner\": \"冯立彬\"\n\t}]\n}"
            }
          }
        }
      }
    },
    "/app/monitor/config/list": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "应用监控配置列表",
        "description": "```\n{\n  \"msg\": \"成功\",\n  \"result\": 1,\n  \"data\": {    \"records\": [\n      {\n        \"id\": \"1\",\n        \"appName\": \"雷神网关应用erp-gateway异常监控\",\n        \"appOwner\": \"\",\n        \"esIndex\": \"erp-gateway-*\",\n        \"ddRobotMobile\": \"15338887312\",\n        \"createTime\": \"2020-12-16 13:58:03\",\n        \"createUserName\": null,\n        \"updateTime\": \"2020-12-18 18:04:16\",\n        \"updateUserName\": null,\n        \"alarmThreshold\":1,\n        \"phoneAlarmThreshold\":10,\n        \"executeCycle\": 10,\n        \"deptId\":1,\n        \"deptName\":\"雷神项目部\",\n        \"testerName\":\"测试\",\n        \"testerPhone\":\"13888888888\"\n      }    ],\n    \"total\": 1,\n    \"size\": 10,\n    \"current\": 1,\n    \"searchCount\": true,\n    \"pages\": 1\n  }}\n```",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "configId": {
                  "type": "integer",
                  "description": "监控项目ID"
                },
                "alarmMobile": {
                  "type": "string",
                  "description": "告警手机号码"
                },
                "esIndex": {
                  "type": "string",
                  "description": "ES索引"
                },
                "pageNum": {
                  "type": "integer",
                  "mock": {
                    "mock": "1"
                  },
                  "description": "第几页"
                },
                "pageSize": {
                  "type": "integer",
                  "mock": {
                    "mock": "10"
                  },
                  "description": "每页大小"
                },
                "appOwner": {
                  "type": "string",
                  "description": "应用负责人的名称",
                  "mock": {
                    "mock": "冯立彬"
                  }
                },
                "deptId": {
                  "type": "integer",
                  "mock": {
                    "mock": "1"
                  },
                  "description": "应用归属项目部的ＩＤ"
                }
              },
              "required": [
                "pageNum",
                "pageSize"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {}
            }
          }
        }
      }
    },
    "/app/monitor/config/delete": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "删除应用配置",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "待删除的配置id",
            "type": "string"
          },
          {
            "name": "id",
            "in": "formData",
            "required": true,
            "description": "待删除的配置id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\"type\":\"object\",\"title\":\"empty object\",\"properties\":{}}"
            }
          }
        }
      }
    },
    "/app/monitor/config/save": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "新增应用配置",
        "description": "",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "raw",
            "in": "body",
            "description": "raw paramter",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n  \"appName\":\"crm异常监控\",\n  \"appOwner\":\"老吴\",\n  \"esIndex\":\"ld-erp-crm*\",\n  \"ddWebhookAccessToken\":\"d361cf7a46e47588395372602549149e243ac7fa8f1cc84e1e5eea3c339a7a47\",\n  \"esParam\":\"{\\\"query\\\":{\\\"bool\\\":{\\\"filter\\\":[{\\\"term\\\":{\\\"message\\\":\\\"error\\\"}},{\\\"range\\\":{\\\"index_utc_date\\\":{\\\"gte\\\":\\\"{startTime}\\\",\\\"lt\\\":\\\"{endTime}\\\"}}}],\\\"must_not\\\":[{\\\"term\\\":{\\\"log.file.path\\\":\\\"info.log\\\"}}]}}}\",\n  \"ddRobotMobile\":\"15338887312\",\n  \"alarmThreshold\":1,\n  \"phoneAlarmThreshold\":10,\n  \"executeCycle\": 10,\n  \"deptId\":1,\n  \"testerName\":\"测试\",\n  \"testerPhone\":\"13888888888\"\n}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n    \"msg\": \"成功\",\n    \"result\": 1,\n    \"data\": true\n}"
            }
          }
        }
      }
    },
...
  }
}